# bank-lending-frontend/src/app/services/models/client-information-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ClientInformationResponse {
  balance?: number;
  firstName?: string;
  lastName?: string;
  passportSerialNumber?: string;
  previousLoansInfo?: string;
  surName?: string;
}