# bank-lending-frontend/src/app/services/models/credit-condition-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreditConditionRequest {
  amount?: number;
  id?: number;
  interestRate?: number;
  monthlyPayment?: number;
  term?: number;
}