# bank-lending-frontend/src/app/services/models/credit-condition-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreditConditionResponse {
  id?: number;
  interestRate?: number;
  monthlyPayment?: number;
  productName?: string;
  term?: number;
}