# bank-lending-frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "contact": {
      "name": "Timur Sharf",
      "url": "https://github.com/xzcPozer",
      "email": "timurkasharf@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/credit-condition/send/final-condition/{userId}": {
      "post": {
        "tags": [
          "CreditCondition"
        ],
        "operationId": "sendBestConditionForClient",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditConditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-condition/recalculation-monthly-payment": {
      "post": {
        "tags": [
          "CreditCondition"
        ],
        "operationId": "recalculationOfMonthlyPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number",
                  "format": "double"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/change/password": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "login_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/credit-condition/condition/by/condition-name": {
      "get": {
        "tags": [
          "CreditCondition"
        ],
        "operationId": "getConditionByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditConditionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/credit-condition/best-condition/{userId}": {
      "get": {
        "tags": [
          "CreditCondition"
        ],
        "operationId": "getBestConditionForUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditConditionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/client-information/client/by/serial-number": {
      "get": {
        "tags": [
          "ClientInformation"
        ],
        "operationId": "getClientInfoBySerialNumber",
        "parameters": [
          {
            "name": "serialNum",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientInformationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/client-information/all-client": {
      "get": {
        "tags": [
          "ClientInformation"
        ],
        "operationId": "getAllClientInfo",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseClientInformationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/client-information/all-client/by/sort": {
      "get": {
        "tags": [
          "ClientInformation"
        ],
        "operationId": "getAllClientInfoBySort",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseClientInformationResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreditConditionRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "term": {
            "type": "integer",
            "format": "int32"
          },
          "interestRate": {
            "type": "number",
            "format": "double"
          },
          "monthlyPayment": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PaymentCalculationRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "term": {
            "type": "integer",
            "format": "int32"
          },
          "interestRate": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "LoginChangeRequest": {
        "required": [
          "confirmNewPassword",
          "email",
          "newPassword",
          "oldPassword"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "confirmNewPassword": {
            "type": "string"
          }
        }
      },
...
      "CreditConditionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "interestRate": {
            "type": "number",
            "format": "double"
          },
          "term": {
            "type": "integer",
            "format": "int32"
          },
          "monthlyPayment": {
            "type": "number",
            "format": "double"
          },
          "productName": {
            "type": "string"
          }
        }
      },
      "ClientInformationResponse": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "passportSerialNumber": {
            "type": "string"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "previousLoansInfo": {
            "type": "string"
          }
        }
      },
      "PageResponseClientInformationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientInformationResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth with bearer token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}